refactor: create azure module
