+++ cbsurge/exposure/population/download_worldpop.py
@@ -9,52 +9,15 @@ from html.parser import HTMLParser
 import aiofiles
 import httpx
 
-from azure.storage.blob.aio import BlobServiceClient
 from tqdm.asyncio import tqdm
 
+from cbsurge.azure_upload import AzStorageManager
 
 logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
 logging.getLogger("azure").setLevel(logging.WARNING)
 
 CONTAINER_NAME = "stacdata"
 AZ_ROOT_FILE_PATH = "worldpop"
-
-class AzStorageManager:
-    def __init__(self, conn_str):
-        logging.info("Initializing Azure Storage Manager")
-        self.conn_str = conn_str
-        self.blob_service_client = BlobServiceClient.from_connection_string(conn_str)
-        self.container_client = self.blob_service_client.get_container_client(CONTAINER_NAME)
-
-    async def upload_file(self, file_path=None, blob_name=None):
-        """
-        Upload a file to Azure Blob Storage.
-        Args:
-            file_path: (str) The path to the file to upload
-            blob_name: (str) The name of the blob to create
-
-        Returns:
-
-        """
-        logging.info("Starting upload for blob: %s", blob_name)
-        blob_client = self.container_client.get_blob_client(blob_name)
-
-        # Get the file size for progress tracking
-        file_size = os.path.getsize(file_path)
-
-        async def __progress(current, total):
-            tqdm(total=total, unit="B", unit_scale=True, desc=f"Uploading {blob_name}").update(current)
-
-        async with aiofiles.open(file_path, "rb") as data:
-            await blob_client.upload_blob(data, overwrite=True, max_concurrency=4, blob_type="BlockBlob", length=file_size, progress_hook=__progress)
-
-        logging.info("Upload completed for blob: %s", blob_name)
-        return blob_client.url
-
-    async def close(self):
-        logging.info("Closing Azure Storage Manager")
-        await self.blob_service_client.close()
-        await self.container_client.close()
 
 
 class LinkExtractor(HTMLParser):
@@ -110,7 +73,7 @@ async def get_available_data(country_code=None, year="2020"):
     Returns:
 
     """
-    url = "https://hub.worldpop.org/rest/data/age_structures/ascicua_2020"
+    url = f"https://hub.worldpop.org/rest/data/age_structures/ascicua_{year}"
     logging.info("Fetching available countries from: %s", url)
     try:
         async with httpx.AsyncClient() as client:
@@ -252,7 +215,8 @@ async def process_single_file(file_url=None, storage_manager=None, country_code=
                                         logging.info("Download path does not exist. Creating download path: %s", download_path)
                                         os.makedirs(download_path, exist_ok=True)
                                     logging.info("Copying COG file locally: %s", cog_path)
-                                    shutil.move(cog_path, f"{download_path}/{year}/{country_code}{file_name}")
+                                    os.makedirs(f"{download_path}/{year}/{country_code}", exist_ok=True)
+                                    shutil.move(cog_path, f"{download_path}/{year}/{country_code}/{file_name}")
                                     logging.info("Successfully copied COG file: %s", f"{download_path}/{year}/{country_code}/{file_name}")
                                 except Exception as e:
                                     raise Exception(f"Error copying COG file: {e}")
